@@ -16,31 +16,18 @@
     "vue-devtools": "vue-devtools",
     "npm-run-all": "npm-run-all",
     "rimraf": "rimraf",
-    "concurrently": "concurrently",
     "webpack-cli": "webpack-cli",
     "compile": "npm-run-all webpack-cli mock_for_oauth2orize",
     "dev:webpack:watch": "webpack-cli --watch",
     "dev:electron": "electron .",
     "cleanup": "rimraf ./dist",
     "mock_for_oauth2orize": "mkdir -p ./dist/grant && mkdir -p ./dist/exchange",
-    "dev": "npm-run-all cleanup compile \"concurrently \"npm:dev:*\"\"",
-    "storybook:build": "vue-cli-service storybook:build -c config/storybook",
-    "storybook:serve": "vue-cli-service storybook:serve -p 6006 -c config/storybook",
     "test": "jest"
   },
-  "dependencies": {
-    "@ant-design/icons-vue": "^5.1.7",
-    "@matrixai/polykey": "^0.0.23",
-    "ant-design-vue": "^2.0.0-beta.15",
-    "balm-ui": "^9.0.0-rc.4",
-    "electron": "^9.3.5",
-    "jdenticon": "^3.1.0",
-    "tailwindcss": "^2.0.0",
-    "vue": "^3.0.2",
-    "vue-router": "^4.0.0-0",
-    "vuex": "^4.0.0-0"
-  },
+  "dependencies": {},
   "devDependencies": {
+    "@ant-design/icons-vue": "^5.1.9",
+    "@matrixai/polykey": "0.0.23",
     "@types/copy-webpack-plugin": "^6.0.0",
     "@types/jest": "^25.0.0",
     "@types/node": "^13.13.1",
@@ -55,10 +42,11 @@
     "@vue/cli-plugin-vuex": "~4.5.0",
     "@vue/devtools": "^6.0.0-beta.2",
     "@vue/test-utils": "^2.0.0-beta.5",
-    "concurrently": "^5.3.0",
+    "ant-design-vue": "^2.0.0-rc.8",
     "copy-webpack-plugin": "^6.3.0",
     "cross-env": "^7.0.2",
     "css-loader": "^4.3.0",
+    "electron": "^9.4.0",
     "electron-builder": "^22.9.1",
     "electron-packager": "^15.0.0",
     "eslint": "^7.6.0",
@@ -70,6 +58,7 @@
     "eslint-plugin-vue": "^7.0.0-0",
     "file-loader": "^6.1.0",
     "html-webpack-plugin": "^4.5.0",
+    "jdenticon": "^3.1.0",
     "jest": "^25.0.0",
     "less-loader": "^7.1.0",
     "mini-css-extract-plugin": "^1.2.0",
@@ -78,17 +67,20 @@
     "postcss-loader": "^4.0.4",
     "prettier": "^1.19.1",
     "rimraf": "^3.0.2",
-    "storybook": "^5.3.19",
     "style-loader": "^2.0.0",
+    "tailwindcss": "^2.0.2",
     "ts-loader": "^8.0.7",
     "ts-node": "^8.8.2",
     "tsconfig-paths-webpack-plugin": "^3.3.0",
     "typescript": "~3.9.3",
     "unzip": "^0.1.11",
+    "vue": "^3.0.5",
     "vue-jest": "^5.0.0-0",
     "vue-loader": "^16.0.0-beta.5",
+    "vue-router": "^4.0.3",
     "vue-style-loader": "^4.1.2",
     "vue-svg-loader": "^0.17.0-beta.2",
+    "vuex": "^4.0.0-rc.2",
     "webpack": "^5.2.0",
     "webpack-cli": "^4.2.0",
     "webpackbar": "^4.0.0"
