@@ -1,4 +1,4 @@
-import * as grpc from '@grpc/grpc-js'
+import * as grpc from '@grpc/grpc-js';
 
 function promisifyGrpc<t1, t2>(
   fn: (request: t1, callback: (error: grpc.ServiceError | null, response: t2) => void) => any
@@ -7,13 +7,13 @@ function promisifyGrpc<t1, t2>(
     return new Promise<t2>((resolve, reject) => {
       function customCallback(error: grpc.ServiceError | null, response: t2) {
         if (error) {
-          return reject(error)
+          return reject(error);
         }
-        return resolve(response)
+        return resolve(response);
       }
-      fn(request, customCallback)
-    })
-  }
+      fn(request, customCallback);
+    });
+  };
 }
 
-export { promisifyGrpc }
+export { promisifyGrpc };
