@@ -1,9 +1,9 @@
 <template>
   <div class="w-72 mb-1">
-    <a-progress :showInfo="false" :percent="30" />
+    <a-progress :showInfo="false" :percent="30" strokeColor="#116466" />
   </div>
-  <div class="font-roboto text-xs w-72 bg-gray-50 p-5 h-15">
-    <div class="overflow-y-auto h-8">
+  <div class="font-robotomono text-xs w-72 text-content5 p-5 h-15">
+    <div class="overflow-y-auto h-8 ">
       <p class="opacity-50 -mb-0.5">Initializing...</p>
       <p class="opacity-50 -mb-0.5">Initializing Polykey...</p>
       <p class="opacity-50 -mb-0.5">Intalling Agent...</p>
