@@ -4,14 +4,14 @@
   </div>
 </template>
 <script>
-import Bell from '@renderer/assets/bell.svg'
-import Badge from '@renderer/assets/badge.svg'
+import Bell from '@renderer/assets/bell.svg';
+import Badge from '@renderer/assets/badge.svg';
 
-import { defineComponent } from 'vue'
+import { defineComponent } from 'vue';
 export default defineComponent({
   components: {
     Bell,
     Badge
   }
-})
+});
 </script>
