@@ -5,8 +5,8 @@
   </div>
 </template>
 <script>
-import Search from '@renderer/assets/magnify.svg'
-import { defineComponent } from 'vue'
+import Search from '@renderer/assets/magnify.svg';
+import { defineComponent } from 'vue';
 export default defineComponent({
   props: {
     type: {
@@ -15,22 +15,22 @@ export default defineComponent({
     }
   },
   data() {
-    let defaultStyle = 'rounded-md bg-grey2 border border-content4 pl-8 pr-2'
-    let defaultIconStyle = 'absolute top-0 ml-2'
+    let defaultStyle = 'rounded-md bg-grey2 border border-content4 pl-8 pr-2';
+    let defaultIconStyle = 'absolute top-0 ml-2';
 
-    let height = 'h-search'
-    let iconMargin = 'mt-1'
+    let height = 'h-search';
+    let iconMargin = 'mt-1';
     if (this.type && this.type == 'large') {
-      height = 'h-large-input'
-      iconMargin = 'mt-2'
+      height = 'h-large-input';
+      iconMargin = 'mt-2';
     }
     return {
       inputClass: `${defaultStyle} ${height}`,
-      iconClass : `${defaultIconStyle} ${iconMargin}`
-    }
+      iconClass: `${defaultIconStyle} ${iconMargin}`
+    };
   },
   components: {
     Search
   }
-})
+});
 </script>
