@@ -4,11 +4,11 @@
   </div>
 </template>
 <script>
-import Cog from '@renderer/assets/cog.svg'
-import { defineComponent } from 'vue'
+import Cog from '@renderer/assets/cog.svg';
+import { defineComponent } from 'vue';
 export default defineComponent({
   components: {
     Cog
   }
-})
+});
 </script>
