@@ -1,5 +1,5 @@
-import PolykeyClient from '@renderer/resources/PolykeyClient'
-import * as pb from '@matrixai/polykey/proto/compiled/Agent_pb'
+import PolykeyClient from '@renderer/resources/client';
+import * as pb from '@matrixai/polykey/proto/compiled/Agent_pb';
 
 export default {
   namespaced: true,
@@ -14,45 +14,45 @@ export default {
   },
   actions: {
     loadPeerIds: async function({ commit }) {
-      const peerIds = await PolykeyClient.ListPeers()
-      commit('setPeerNames', peerIds)
+      const peerIds = await PolykeyClient.ListPeers();
+      commit('setPeerNames', peerIds);
     },
     selectPeerId: async function({ commit }, peerId: string) {
-      const peerInfo = await PolykeyClient.GetPeerInfo(peerId)
-      commit('setSelectedPeer', { peerId, peerInfo })
+      const peerInfo = await PolykeyClient.GetPeerInfo(peerId);
+      commit('setSelectedPeer', { peerId, peerInfo });
 
       PolykeyClient.ScanVaultNames(peerId)
         .then((scannedVaultNames: string[]) => {
-          commit('setScannedVaultNames', scannedVaultNames)
+          commit('setScannedVaultNames', scannedVaultNames);
         })
-        .catch(e => console.log(`error scanning vault names: ${e.message}`))
+        .catch(e => console.log(`error scanning vault names: ${e.message}`));
     },
     pullVault: async ({ commit }, { peerId, vaultName }: { peerId: string; vaultName: string }) => {
       try {
         await PolykeyClient.PullVault({
           publicKey: peerId,
           vaultName: vaultName
-        })
+        });
       } catch (error) {
-        console.log(error)
+        console.log(error);
       }
     }
   },
   mutations: {
     setPeerNames: (state, peerIds: string[]) => {
-      state.peerIds = peerIds
+      state.peerIds = peerIds;
     },
     setSelectedPeer: (state, { peerId, peerInfo }: { peerId: string; peerInfo: pb.PeerInfoMessage.AsObject }) => {
-      state.selectedPeerId = peerId
-      state.publicKey = peerInfo.publicKey ?? ''
-      state.rootCertificate = peerInfo.rootCertificate ?? ''
-      state.peerAddress = peerInfo.peerAddress ?? ''
-      state.apiAddress = peerInfo.apiAddress ?? ''
-      state.scannedVaultNames = []
+      state.selectedPeerId = peerId;
+      state.publicKey = peerInfo.publicKey ?? '';
+      state.rootCertificate = peerInfo.rootCertificate ?? '';
+      state.peerAddress = peerInfo.peerAddress ?? '';
+      state.apiAddress = peerInfo.apiAddress ?? '';
+      state.scannedVaultNames = [];
     },
     setScannedVaultNames: (state, scannedVaultNames: string[]) => {
-      state.scannedVaultNames = scannedVaultNames
+      state.scannedVaultNames = scannedVaultNames;
     }
   },
   getters: {}
-}
+};
