@@ -1,38 +1,73 @@
-import PolykeyClient from '@renderer/resources/PolykeyClient'
+import PolykeyClient from '@renderer/resources/client';
+import { makeIdentifiers } from '@renderer/store/utils';
+
+const [actionsInt, actionsExt] = makeIdentifiers('Vaults', [
+  'NewVault',
+  'SearchFilter',
+  'LoadVaultNames',
+  'SelectVault',
+  'DeleteVault'
+]);
+
+const enum mutations {
+  SetVaultNames = 'SetVaultNames',
+  SetSelectedVault = 'SetSelectedVault',
+  SetSearchFilter = 'SetSearchFilter'
+}
+
+type State = {
+  vaultNames: Array<string>;
+  selectedVaultName: string;
+  searchFilter: string;
+};
+
+const state: State = {
+  vaultNames: [],
+  selectedVaultName: '',
+  searchFilter: ''
+};
+
+export { actionsExt as actions };
 
 export default {
   namespaced: true,
-  state: {
-    vaultNames: [],
-    selectedVaultName: ''
-  },
+  state,
   actions: {
-    loadVaultNames: async function({ commit }, selectFirstVault = false) {
-      const vaultNames = await PolykeyClient.ListVaults()
-      commit('setVaultNames', vaultNames)
-      console.log(selectFirstVault)
+    async [actionsInt.NewVault]({ commit }, vaultName: string) {
+      await PolykeyClient.NewVault(vaultName);
+      const vaultNames = await PolykeyClient.ListVaults();
+      commit(mutations.SetVaultNames, vaultNames);
+    },
 
-      if (selectFirstVault && vaultNames.length !== 0) {
-        console.log(vaultNames[0])
-        commit('setSelectedVault', vaultNames[0])
-      }
+    async [actionsInt.SearchFilter]({ commit }, filter: string) {
+      commit(mutations.SetSearchFilter, filter);
     },
-    selectVault: async function({ commit }, vaultName) {
-      commit('setSelectedVault', vaultName)
+
+    async [actionsInt.LoadVaultNames]({ commit }) {
+      const vaultNames = await PolykeyClient.ListVaults();
+      commit(mutations.SetVaultNames, vaultNames);
     },
-    deleteVault: async function({ commit }, vaultName) {
-      await PolykeyClient.DeleteVault(vaultName)
-      const vaultNames = await PolykeyClient.ListVaults()
-      commit('setVaultNames', vaultNames)
+
+    async [actionsInt.SelectVault]({ commit }, vaultName: string) {
+      commit(mutations.SetSelectedVault, vaultName);
+    },
+
+    async [actionsInt.DeleteVault]({ commit }, vaultName: string) {
+      await PolykeyClient.DeleteVault(vaultName);
+      const vaultNames = await PolykeyClient.ListVaults();
+      commit(mutations.SetVaultNames, vaultNames);
     }
   },
   mutations: {
-    setVaultNames: function(state, vaultNames) {
-      state.vaultNames = vaultNames
+    [mutations.SetVaultNames]: function(state: State, vaultNames: Array<string>) {
+      state.vaultNames = vaultNames;
     },
-    setSelectedVault: function(state, vaultName) {
-      state.selectedVaultName = vaultName
+    [mutations.SetSelectedVault]: function(state: State, vaultName: string) {
+      state.selectedVaultName = vaultName;
+    },
+    [mutations.SetSearchFilter]: function(state: State, filter: string) {
+      state.searchFilter = filter;
     }
   },
   getters: {}
-}
+};
